[chore] iarcGroup 파일 분리
